requirements.txt
-e git+https://github.com/wolru/pytrends.git@215a524360158d4e2259e9772ab2fe89175eebf3#egg=pytrends
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langchain-openai==0.0.5
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==1.25.4